拆分语音检查功能
